+++ NEWS.md
@@ -1,5 +1,6 @@
 # ggpcp 0.2.9000
 
+* `plotly` dependency for interactive use
 * update website for better accessibility
 * add JCGS paper as citation